Move vulnerability tests into new structure
